Align `ApproxEq` trait with `allclose` in PyTorch / NumPy
Align the way approximate equality is determined to match the Python libraries
which are a reference for most of the operations implemented here. This means
taking into account both the absolute and relative differences of the values
being compared, and also using the same default tolerances for float values.

Taking into account the relative difference allowed for custom looser tolerances
to be dropped in a couple of cases.