Fallback to bilinear resizing with a warning if bicubic is requested
Add a mechanism for model conversion to substitute certain operator attributes
with others, but with a warning, if such conversions are likely to be "safe" in
the sense they will not affect the model's ability to run and produce useful
results, but may only have a small impact on accuracy.

The initial use is downgrading unsupported bicubic resizing to bilinear. This
allows the YOLOS model family [1] to be used with the DETR example.

[1] https://huggingface.co/hustvl/yolos-small