Add `NdTensorBase::{into_dyn, assign_array, to_array}`
Add three methods that were useful when working on the YOLO example.

 - `into_dyn` is the moving counterpart to `as_dyn`
 - `to_array` is a shorthand for `get_array([0], 0)` for extracting a
   dimension of statically known size into an array
   `assign_array` is the mutating counterpart to `to_array`

In writing the tests I found an issue where `NdTensorBase::transposed` consumed
its input when it should return a view like `NdView::permuted`.  This method
could just be part of the `NdView` trait but it would have to be generalized to
work with ND views, eg. by just reversing the order of dimensions.