Fix broken Softmax and LogSoftmax operators when applied to an inner axis
The Softmax and LogSoftmax are supposed to be applied to 1D lanes of the
input. The previous logic worked in the common case when the operation
was applied to the rightmost axis, but was being applied to the wrong
elements otherwise.

Since Softmax is a performance-sensitive operation in many models, this fix
keeps the inner loop operating over contiguous slices but temporarily swaps
dimensions if the operation is applied to an inner dimension.

There were tests for this, but they were wrong. The expected results have been
recomputed using PyTorch's `torch.Tensor.{softmax, log_softmax}` methods.