Make `{NdTensor, Tensor}::{data, data_mut}` return `None` if non-contiguous
If the tensor is non-contiguous, `data_mut` can return slices which overlap with
other views, violating Rust's rules on mutable references being unique. Also
with both `data` and `data_mut`, if callers failed to check if the tensor was
contiguous, they could silently end up operating on the wrong elements.

To address both of these issues, make `data` and `data_mut` return `None` if
the view is not contiguous. There are some cases where callers need to be
able to access the underlying data of a view and handle only accessing valid
offsets themsleves. For this purpose add unsafe `_unchecked` variants of the
methods.

As part of this, make the `is_contiguous` helper generic so that it can be
optimized for the statically-known size of NdTensor shape/stride arrays.