Add fast path for broadcasting in non-commutative binary ops like Sub
The in-place binary op inner loop had a fast path for when the LHS and RHS where
contiguous and broadcasting the RHS via cycles + repeats is possible. Add the
same optimization for non-in-place operators, and remove the now-unnecessary
code to convert non-in-place commutative binary ops (eg. Add, Mul) to in-place
ones. Also add an extra test for broadcasting that explicitly covers this path.

In the process the `Default` trait was added as a bound because it is needed for
`Tensor::zeros`. This could have been avoided with some more unsafe code to
populate an uninitialized vector, however I saw no performance benefit.

On my system this reduces time in `Sub` ops in MobileViT from ~5ms to ~0.4ms.