Switch to `image::imageops::resize` for resizing in the imagenet demo
The `imageops::resize` function produces better quality outputs when the output
is small, and has been significantly downscaled from the input, because it does
antialiasing. The the implementation in the `Resize` operator does not.

This improves the alignment between prediction outputs of the
`tools/export-timm-model.py` script and the imagenet demo. The results are still
not perfectly identical because the transforms don't exactly match.