Optimize index iteration over dynamic-rank tensors
The generic code in `<Indices<A> as Iterator>::next` is much faster for
fixed-sized arrays than `SmallVec<A>`.

Most dynamic-rank tensors only have a small number of dims, so modify DynIndices
to be a wrapper which switches between `Indices<[usize; 4]>` and
`Indices<SmallVec<...>>` depending on the tensor rank.

Add a micro-benchmark which runs ~2.2x faster if DynIndices uses the fast path
for tensors of rank <= 4 than the slower path, which is the same as the old
code. The dynamic case is still ~2-3x slower than the static case for the same
shape.

The code in `Indices::from_start_and_end` which tests whether a range of indices
is empty was updated so that it works if an empty range is padded with `0..1`
(eg. `[0..0]` => `[0..1, 0..1, 0..0]`). In this case the start/end indices are
different, but the padded range is still empty.