Optimize GatherElements - round 2
Rework GatherElements so that it does not have to explicitly construct indices
into the `input` tensor and use them to fetch elements. Instead it is possible
to do implicit indexing by keeping track of the iteration offset and
manipulating it to construct the offset into the input data. This reduces time
in GatherElements in DeBERTa Large from ~110ms to ~80ms. As a bonus we now no
longer need special cases depending on the input tensor.