Optimize Slice operations that involve copying data
In all cases except where the step size is negative, this operation can create a
view with `Tensor::try_slice` and then create a contiguous owned output using
`Tensor::to_tensor`. This is faster than using `Tensor::slice_iter` because
`to_tensor` has optimizations for the cases where the sliced view is contiguous
(the data is just memcpy-d) or has <= 4 dims.

Tested with the Jina embeddings model [1], this reduces time in Slice ops by ~3x
(~18ms => ~6ms for an input sequence length of 128).

[1] jinaai/jina-embeddings-v2-base-en