Extract reusable parts of WordPiece tokenization
Extract the parts of the WordPiece tokenizer that are not specific to that
tokenization method into a shared `Tokenizer` struct. This struct wraps an
`Encoder` trait that is implemented by WordPiece and in future other models
(Bpe, Unigram etc.) and adds the logic for inserting special tokens, chunking
the output etc.

All of the tests still use the WordPiece encoder, but those tests which are
mostly concerned with the shared functionality have been moved to tokenizers.rs.