Use a separate associated type for SIMD masks
For WebAssembly and AVX2, the `Self` type can also be used as the SIMD mask
type. However for i32/f32 and ARM the code had to convert to/from the actual
mask type (bool and uint32_t respectively) to the `Self` type.  Remove this
unnecessary conversion and make the APIs clearer by introducing a separate
`Mask` associated type for SIMD vectors.