Optimize depthwise convolution
Add a series of optimizations which reduce the overhead around the inner `axpy`
operation. The benefit is inversely proportional to the input width (ie. has
more benefit for small spatial dims and many channels), but always positive.

 - Replace view slicing with manual slicing of the underlying buffer.
   View slicing could be optimized for the special case of slicing with
   a single index, but is still going to be more expensive than doing it
   manually.

 - Inline `add_scaled_vector`. This reduces function call overhead in the inner
   loop and allows codegen to specialize for `dest_stride` being a constant.

 - Simplify `add_scaled_vector` using `unroll_loop!`. This had no performance
   impact, but it's shorter.

 - Precalculate input/output column ranges for each kernel position outside the
   conv loops. The precalculated ranges are stored in a `SmallVec` sized to
   avoid allocation for common kernel sizes.

In the depthwise conv benchmark based on MobileNetV2, which uses small 3x3
kernels of stride 1 or 2, these changes add up to an improvement of ~35%.