Add `Layout::offset` method and remove redundant bounds on `Layout::Index`
A layout has two functions: 1) defining the shape of a tensor and 2) the mapping
between indices and offsets. Given (2), a method to map between index and offset
is something the trait should obviously have, and enables writing more code that
works generically with different layouts. Adding this required removing some
bounds from `Layout::Index` which are apparently not needed any longer.