Initial implementation of unified static/dynamic rank tensor
Add a new tensor base type in `rten_tensor::unified_tensor` which is generic
over the layout and can represent either a static or dynamic rank tensor.
Unifying the implementation will help to avoid unintended API differences
between the two.

The API is largely compatible with the existing tensor types and the same set of
aliases are provided (`NdTensor*` for static rank, `Tensor*` for dynamic rank).