Optimize index iteration using `zip` + `rev`.
Replace the `Iterator::next` impl for index iterators with a version adapted
from ndarray's `Dimension::next_for` implementation. With the previous code,
most of the time in iteration was spent moving generated index values between
memory and registers. With the revised code, the compiler is able to translate
iteration into a set of nested loops without the mem <-> reg moves.

Using the `bench_indices` test, which iterates 100 times over a `[16,
128, 128]` shape:

Before:
  DynIndices stepped 26214400 times in 245 ms
  NdIndices stepped 26214400 times in 57 ms

After:
  DynIndices stepped 26214400 times in 73 ms
  NdIndices stepped 26214400 times in 11 ms