Optimize `col2im` by assuming input is contiguous
This enables skipping offset calculations for the `columns` input and instead we
just iterate over the underlying data directly. The downside is that if the
input is not contiguous, there will be an expensive copy of the input first.

This improves the `bench_col2im` benchmark from ~140ms to ~115ms.