Optimize `find_contours` for large images by using `i8` working space
Use an `i8` rather than `i32` matrix for the working space in `find_contours`.
This reduces memory usage and speeds up searches in larger images. We can get
away with `i8` because we only need to use 4 distinct labels for pixels, as we
don't implement the part of the original algorithm that determined the
hierarchical relationship between orders.