Add alternative strategy for batched matrix multiplication
Batched matrix multiplication was handled by prepacking one or neither of
the inputs, depending on how often each is re-used, and then performing
one `gemm` call per matrix in the output shape.

This can be inefficient the LHS input has a small number of rows. For example in
[1], the LHS / "A" input is a row vector. In the case where the "A" input is a
batch and the "B" input is a single matrix, the "A" input can be reshaped so a
single gemm call can be used, with the output reshaped afterwards to restore the
batch dimensions.

Implement this alternate approach and add a simple benchmark for batched matmul.

[1] https://github.com/robertknight/rten/issues/50