Use `ubyte` for OperatorType enum
This increases the number of operator types that can be declared without
resorting to a breaking change or negative enum values.

The current values are all <= 127 and the new type has the same size and
alignment as the previous type, so this change is backwards compatible.

See also https://github.com/robertknight/rten/issues/54