Add optimized GEMM implementation for vector-matrix multiplications
When the LHS / "A" matrix argument to a matrix multiplication is a row vector,
use an optimized vector-matrix ("gemv") function which avoids the overhead of
packing, as otherwise packing costs consume most of the execution time.

Vector-matrix multiplications are common in autoregressive transformers,
including LLMs, for example.

Rather than implement the same algorithm separately for each architecture, the
generic SIMD wrappers in rten-vecmath are reused.

Tested on an Intel i5-1038NG7, this improves the performance of the
vector-matrix product in the `bench_gemm` benchmark by ~4x (5 -> 20 GFLOPS).