Replace architecture-specific GEMM microkernel with a generic implementation
The GEMM microkernel was essentially the same for Intel, Arm and
non-architecture specific, except for the instructions used. Having a generic
implementation makes it easier to change the code or bring up a baseline for
new architecture.

One small change is that the unroll factor is now 4 everywhere. Previously there
was no explicit unrolling for Intel and 8x for Arm. The `unroll_loop!` macro
doesn't guarantee unrolling, it only generates code which is very likely to be
unrolled, as it uses a small static iteration count. Using `unroll_loop!` with
the previous 8x factor in the generic GEMM microkernel resulted in the loop no
longer being unrolled under Arm. Using a small unroll factor (4x) or replacing
`unroll_loop!` with an explicit 8x-unroll resolved the issue. `unroll_loop!`
with the smaller factor worked better for both Intel and Arm, so that's what I'm
using.