Align name of `Forward` enum variant for RNN ops with ONNX attribute value
Fix an error when converting LSTM / GRU operators with a direction attribute of
"forward", or with no direction specified, as "forward" is the default. The
enum name was previously `Forwards`, so the conversion code expected an
attribute value of `forwards`.

The problem was spotted when converting the `decoder_iter.onnx` model from
https://github.com/xd009642/xd-tts/tree/master/models/tacotron2.