Refactor GEMM kernel construction
Previously it was the caller's responsibility to ensure that a kernel was
supported on the current system using `Kernel::supported` before calling unsafe
methods on a kernel that require this. Flip this around so that it is the
kernel type's responsibility to only allow it to be constructed if the
instructions are supported, or the kernel implements a fallback internally.

This allows all Kernel methods to assume that the instruction set if supported
if they take `&self` as a parameter.