Add node name to timing reports
For each distinct input shape in the timing report tables produced by
`RTEN_TIMING=by-shape=1`, include the name of an example operator node which had
this shape during the run.

This makes it easier to go from the timing stats in `RTEN_TIMING` output to
corresponding model nodes in Netron etc. This helps with finding higher-level
optimizations.