Add API for partial evaluation of a model graph
Add `Model::partial_run` API to optimize use cases where a model is run
repeatedly in a loop, and some of the inputs are constant across several
invocations of the model. In this scenario the `partial_run` API can be called
once with the inputs that are constant over the loop to pre-evaluate the parts
of the graph that depend only on those inputs, and the results can then be
passed to `Model::run` in the loop for faster evaluation of the remainder of the
graph.

Tested with an implementation of Whisper which uses a transformer encoder +
decoder, this reduces the runtime of the model by over 50%, as it turns out that
the most expensive operation in the decoder model is computing cross-attention
keys and values, which depends only on the encoder outputs.