Compute operator inputs in right-to-left order
To improve the chances of being able to run an operator in-place, visit inputs
in right-to-left order during planning. This results in inputs being computed in
this order by the generated plan, which means that the left-most input is more
likely to be available for mutation (ie. not used by other operators) at the
point when the operator is run.

See https://github.com/robertknight/rten/issues/98 for an example.

I haven't proven that this change will pick the optimal plan execution order
that maximizes the number of operations run in place, so consider it a
heuristic.