Generalize `TensorBase::assume_init` to work with views
This enables getting an initialized view of part of a tensor after initializing
it.

Add a restriction that the view must be contiguous, to avoid hazards if not all
of the elements in the underlying storage of a view are addressed by the view.
For example, `tensor.assume_init().non_contiguous_data()` could expose
uninitialized elements as initialized in such a case. The current implementation
checks that the layout is contiguous, but we could weaken this to just require
that there are no un-addressed / unused elements in the storage in future.