Accelerate im2col transform using SIMD
Re-implement the fused im2col transform (`VirtualIm2Col`) used by convolutions,
with SIMD operations. This helps by:

 1. Computing the source offsets, gathering and storing multiple elements at once.
 2. Eliminating unpredictable branches to determine whether an element is in
    the padding region or not. Instead predication is used.
 3. Enabling the compiler to unroll inner loops by providing statically known
    loop counts.

On an Intel i5, this reduced time in im2col operations in Ocrs by about 35%, of
which 10% comes from (3). The remaining time is almost all in copying memory.
It might be possible to reduce that with prefetching.