Refactor and expose kernel type used by `GemmExecutor`
This will be needed for a SIMD im2col operation to select the appropriate
implementation to match the GEMM kernel.

As part of this, rename `KernelHint` => `KernelType` and replace
`KernelHint::Auto` by using an `Option<KernelType>` to indicate whether to use a
specific kernel or the default. This allows the non-optional `KernelType` to
represent the kernel that was chosen.