Avoid redundant zeroing of output buffers for unary ops
For unary op executions that are not in-place, the output buffer does not need
to be zero-initialized.

The softmax implementation was converted to use a possibly-uninitialized output
buffer in this PR, although the `Softmax` operator in the `rten` crate doesn't
benefit yet, as it copies the input and then runs the in-place implementation.