Support AVX-512 for SIMD unary operators, softmax
Support AVX-512 for Exp, Sigmoid, Tanh, Erf, Softmax and other operators which
use SIMD-accelerated implementations in rten-vecmath.

The performance benefit is small because these operations are limited by memory
bandwidth, but doing this revealed some issues with supporting different
instruction sets on the same architecture. Usage of `#[inline]` and
`#[target_feature]` had to be corrected to ensure that SIMD intrinics are
actually inlined reliably. Otherwise the function call overhead negates the
benefits of SIMD entirely.

The `SimdInt` and `SimdFloat` docs have been updated to explain the rules around
inlining and target features.