Implement Softplus op
The ONNX spec states that the input must be 1D tensor, but that looks
like a mistake [1] and the restriction is not implemented here.

The corresponding PyTorch operator [2] reverts to a linear function for inputs
above a threshold, but the ONNX spec makes no mention of this and ONNX Runtime
doesn't, so that is not implemented here either.

[1] https://github.com/onnx/onnx/issues/2563
[2] https://pytorch.org/docs/stable/generated/torch.nn.Softplus.html