Add Piper text-to-speech demo
This uses Piper voice models to convert strings of phonemes to audio in .wav
format. The demo takes phonemes rather than text as input to simplify creating
this first version. The Piper project uses a non-trivial C++ library to handle
text-to-phoneme conversion, which I didn't want to add as a dependency.