Re-implement dispatch of SIMD unary ops using a function rather than macro
Change dispatch of SIMD unary ops to use a function, parametrized by an
`S: SimdUnaryOp` argument, instead of a macro.

The main value of this is that it supports unary operations which have
parameters. This is achieved by adding the parameters to the struct which
implements `SimdUnaryOp`. Functions are also just generally easier to work than
macros.