Avoid redundant zeroing of im2col packing buffer
If the packing buffer needed for im2col operations needs to be resized, avoid
zeroing it, since the packing code will overwrite all elements.

As part of this change, change the caller of the packing functions in `gemm.rs`
to always pass an output slice of exactly the correct size, avoiding the need
for each packing function to zero unused space at the end, in order to fulfil
the contract that the entire slice is initialized after the packing code
returns. Previously gemm.rs would always pass an output slice of the same size
for every block to be packed. However the last row/column block may be smaller,
with unused space at the end.

The `VirtualMatrix` trait is now unsafe due to the added requirement that
implementers must initialize the entire buffer passed to them. An alternative
would be to require implementations to return the input as an `&mut [f32]`
and have the caller verify that it matches.