Optimize `TensorBase::copy_from` for non-contiguous `self`
Replace the slow iterator-based fallback in `TensorBase::copy_from` with a much
faster version which uses nested loops over the innermost 4 dimensions.

This reduces time in the `Pad` operator from ~4.5ms to ~0.8ms in Piper TTS
models. That operator used the fallback because it creates a non-contiguous view
of the non-padded region into which it copies the source.