Revise model loading API for ease of use and efficiency
Previously loading a model involved two steps, each of which copied the entire
model:

 1. Reading the file from disk into a `Vec<u8>`, eg. using
    `std::fs::read`
 2. `Model::load` then received a `&[u8]` slice of this and would copy
    the weights into individual owned tensors.

Usually callers would then not immediately free the buffer loaded in step (1),
so the program would use twice as much memory as needed for the weights.

This commit revises the model loading API to be more convenient to use for
common use cases and more efficient:

 - The `Model::load` API now takes a `Vec<u8>` rather than `&[u8]`.

   `Model` then puts this `Vec<u8>` into a reference-counted container and graph
   nodes reference slices of this. This avoids an expensive copy of all the
   model weights.

 - Add a `Model::load_file` API to simplify the most common use case into
   a single call: `Model::load_file("model.rten")`

 - Add a `Model::load_mmap` API which allows loading a model with zero copying
   of weights by creating a memory-mapped view of the file, which graph nodes
   reference directly. This requires additional dependencies, and so is behind
   an `mmap` feature.

To avoid adding variants of each of these methods that allow customizing the
available operators, a `ModelOptions` struct has been added which is modeled
on `std::fs::OpenOptions` and allows configuring the available operators and
other things that have to be specified before the model is loaded.