Improve how dynamic input sizes are handled in rten CLI
Previously automatic sizes for dynamic axes were determined using simple
heuristics with a value of 256 as a fallback. This could result in very slow
model execution if eg. the batch size was resolved to 256.

Specifying dynamic axis sizes manually was unergonomic because it required
specifying the size of every dimension of an input, including the fixed
dimensions. Also if there were multiple inputs with the same symbolic axis name,
each had to be specified independently.

Change this as follows:

 - Default the size to 1 for any dynamic dimension that is not specified
   manually. This seems more likely to lead to successful and fast model
   execution, if we're going to pick defaults at all.

 - Replace the `--shape` flag with `--size`, which takes `dim_name=size` or
   `input_name.dim_name=size` flags. This means that only dynamic axis sizes
   have to be specified, and if the same symbolic name appears in multiple
   inputs, it only has to be specified once.