Make binary ops over non-contiguous tensors faster
Replace the use of `broadcast_iter` for broadcasting + iteration over
non-contiguous tensors with a more efficient pattern based on iterating over a
statically determined number of inner-most dimensions (4) with nested for-loops,
and a variable number of outer dims. This pattern is already used in other
places (eg. `rten-tensor/src/copy.rs`).