Change Piper demo to accept text instead of phonemes
Add a naive text-to-phoneme converter based on a dictionary built with the
piper-phonemize PyPI package. This converter is naive in the sense that it just
ignores words that don't appear in the dictionary, and doesn't deal with
numbers, punctuation etc. Nevertheless, it can be used to translate simple
user-specified sentences.

The provided `en-us.txt` dictionary was created using the pronounciation-dict.py
script with the `/usr/share/dict/words` file on my system.