Perform constant propagation when running model in rten CLI
When testing `decoder_model.onnx` from
https://huggingface.co/Mozilla/distilvit/tree/main/onnx I found that it was very
slow in RTen compared to ONNX Runtime (>100ms vs 20ms). The main culprit turned
out to be an expensive `Transpose` operator with constant inputs. ORT does
constant propagation as part of its automatic optimizations when the model is
loaded. RTen doesn't have an optimization like this implemented, but by
implementing it "manually" in the CLI, we can see how much benefit this would
provide.