Use reshape strategy in `MatMul` based on batch size rather than dims
The `MatMul` operator has an optimization for the case when the LHS is a batch
and the RHS is a single matrix. In that case the LHS is reshaped into a single
matrix, the matmul is done and the output is reshaped to re-add the batch dims.
This is pointless if the LHS is a batch of size 1 though.

Change this so we decide based on batch sizes rather than dimension count. This
doesn't have a meaningful performance impact as reshaping is cheap, it is just
makes more sense conceptually.