Replace AVX2 gather with scalar loads
On Intel CPUs which have a mitigation applied for the "Downfall" / Gather Data
Sampling (GDS) vulnerability [^1], the performance of the AVX2 gather
instructions are reduced and can be slower than a scalar load.

Tested on the YOLOv8 example on an Ice Lake i5, replacing gathers (used by
im2col) with scalar loads reduces inference time by ~20ms (~180ms => ~160ms).

The mitigation has not been applied for AVX-512 because it didn't improve
performance there.

[^1]: https://en.wikipedia.org/wiki/Downfall_(security_vulnerability)