Re-introduce support for calling `TensorBase::init_from` on mutable views
Update this API following the refactoring of tensor storage to use `ViewMutData`
instead of `&mut [T]` for mutable views. This forces the implementation to
properly handle the case where the `self` view is not contiguous. For this
purpose, introduce a variant of the `copy_into` helper which takes an
uninitialized destination. For `Copy` types this is identical. For non-`Copy`
types, `copy_into_uninit` won't drop any existing elements in the destination
view.