Ensure end offset from `Layout::split` does not exceed storage length
For a 1-sized dimension, it is possible that the stride exceeds the storage
length. In this case `layout.split(axis, 1)` would return storage offsets for
the right part of the split that were out of bounds. Resolve this by using
`layout.min_data_len()` as the start/end offset for the right range, since the
storage must be at least this long.