Support "reflect" padding mode
For efficiency and simplicity, reflection padding is only supported for the last
2 dimensions of the input. Any additional dimensions are treated as batch
dimensions which must have the same size in the input and output.

PyTorch has a constraint that the amount of padding must be less than the size
of a dimension, which is useful since it avoids the need for padding to wrap
around by taking an expensive modulus of the dimension size. NumPy however
allows wrap-around, and the example in the ONNX spec relies on this. In this
implementation, wrap-around is supported but it might be necessary to split the
cases in future for efficiency, to avoid a modulus in the common case where the
padding size is less than the dimension size.