Extract `Generator` and `GeneratorUtils` into `rten-generate` crate
Add a new create which contains utilities for running the decoding loop for
transformer decoder-only and encoder-decoder models. It could also be adapted to
support other kinds of auto-regressive models in future.

The initial implementation assumes the model input and output names match the
conventions of the Hugging Face's Optimum ONNX exporter. This can be made more
flexible in time.