Simplify graph plan representation to `Vec<NodeId>`
Representing execution plans as a `Vec<NodeId>`, rather than `Vec<(NodeId, &'a
OperatorNode)>` allows them to be cloned, cached etc. more easily as it no
longer borrows from the `Graph`. There is an added cost of needing to look up
the `OperatorNode` at each step of execution. This is cheap though since it
just requires a slice index.