Revise BPE tokenizer decoding
`Encoder::get_token_str` was documented as returning the "canonical string" but
didn't make it clear what counted as "canonical". For BPE this returned the
results of decoding the token's bytes to UTF-8, which can fail if the byte
sequence doesn't correspond to a complete UTF-8 sequence.

Revise this so that `get_token_str` returns the string _encoding_ of the token's
bytes, as they appear in eg. `merges.txt` and `tokenizer.json` files, and a new
`Encoder::decode` method instead decodes IDs to readable text, and is what most
consumers will want to use. For the BPE decoder, the `decode` impl concatenates
all the tokens' bytes and decodes the whole buffer at once.