Improve decoding with BPE tokenizers in generator.rs
Previously if a token was encountered which did not correspond to a
complete UTF-8 sequence, it would cause `TextGenerator` to stop
producing output and instead accumulate an increasingly long sequence of
token IDs in its internal buffer. This led to the GPT-2 demo "hanging"
sometimes. The same would also happen for other types of tokenizer
error, eg. if a token ID was missing from the tokenizer's vocabulary.

Address these as follows:

 - Use the new `Encoder::decode` method to decode the token ID sequence
   to a string
 - Only loop internally if decoding fails because of an incomplete UTF-8
   sequence. For all other errors, return them directly.