Pre-allocate space for KV cache in generator
Make growth of the KV cache more efficient by pre-allocating space for the
maximum expected sequence length and passing the KV-cache as an owned tensor
into the model. When the `Concat` operator is used to update the KV-cache, it
will run in-place since the KV-cache input tensor already has enough capacity.
This reduces the cost of updating the KV cache from `O(seq_len)` to `O(1)` at
each step.

For this optimization to work, the model has to satisfy some assumptions:

1. The updated KV cache is produced using `Concat`, not via some other sequence
   of operations.
2. Each KV cache input is passed into `Concat` as the first operand, without any
   earlier operations that would cause it to be copied.
3. The output of each KV cache concatenation is returned from the model as an
   updated KV-cache output, without any operations after the `Concat` that would
   cause it to be copied.

On the GPT-2 demo this reduces generation time significantly for long outputs.
eg. For 300 tokens this reduced generation time from ~28s to ~21s on my laptop.