Support passing owned tensors as model inputs
Change type of model inputs from `Input` to `InputOrOutput`, which is an enum
that is either an `Input` (tensor view) or `Output` (owned tensor).

Supporting owned tensors as inputs is one of the pieces needed to implement
efficient KV-cache extension, by pre-allocating buffers up front and growing
them over the course of a generation loop, instead of copying the whole KV cache
on each step.

In addition to KV-cache optimization, passing an owned tensor as an input
enables other optimizations such as re-using the input tensor's buffer for
intermediate calculations inside the model, saving allocations.

Most code that runs models does not need to change because `InputOrOutput`
implements a set of `From<T>` conversions for the same types of `T` that was
being used to convert tensors to `Input`s.