Skip constant propagation in rten-generate if not required
Since constant propagation is now performed as a graph optimization when the
model is loaded, it only needs to be re-done if additional constants are added.

This change assumes that graph optimizations were enabled when the model was
loaded. If they were disabled, and no other constants were added, this could
lead to expensive re-evaluation of unchanging parts of the graph on each run.