Add downcasting methods for `dyn Operator`
For implementing graph optimizatins it will be useful to be able to extract
operator-specific attributes from operators in a model graph.

If the `trait_upcasting` feature was stabilized, I would have used that via
`(dyn_operator as &dyn Any).downcast_ref::<SomeOperator>()`. Since it isn't,
this approach copies from error downcasting in the standard library.