Make `TensorBase::{inner_iter, inner_iter_mut}` more efficient
The slicing in each call to `next` was slow. Since these iterators yield views
with the same layout each time, we can precompute the layout when the iterator
is constructed. Each call to `next` then just has to compute the storage offset
for the next view.

This optimization is not yet implemented for `inner_iter_dyn`, but could
be in future.