Add V2 RTen model format
Add support for generating `.rten` models in a new V2 format. The changes from
the V1 format are:

 1. Tensor data can be stored separately from the FlatBuffers model, but still
    in the same file. This allows for models that are more than 2GB in size.

 2. The file starts with a 32-byte header that contains file magic,
    version number and offsets and lengths for the sections of the file
    containing the model data

The model FlatBuffers schema is the same as the V1 format, except for additions
to allow constant nodes to refer to data stored outside the FlatBuffers buffer.

The rten-convert tool generates V1 models by default, but will generate V2
models if the `--v2` flag is passed.