Generalize transpose fusion
Revise transpose fusion to be able to support additional operators. In this
initial commit, only `MatMul(Transpose(X), Y)` is fused as before, but this will
make it easier to support other operators in future.

In order to construct a `FusedTranspose` from an arbitrary operator, it was
necessary to support cloning any operator. Since operators are stateless, the
easiest way to do this is to make them reference counted in the graph, then they
can be "cloned" just by creating a new `Arc`.