Adjust `attention_mask` size
In HuggingFace models this input's sequence axis is expected to have the same
size as the sequence length, not the length of the input IDs that are being
provided at the current step. The length was correct for the initial prompt but
wrong for subsequent generation steps. However when only one new token was added
during iterative decoding, the `attention_mask` worked despite being the wrong
size because 1-sized inputs are broadcast by various operators. When appending
multiple tokens after the initial generation however, such as when adding a
tokenized chat message from a user, this broadcasting failed.