Support BPE decoding where token IDs are not 256 + merge rank
In GPT-2 the IDs of tokens are 256 plus the rank of the pair that forms the
token in the BPE merge list. In other tokenizers such as that used by SmolLM [1]
this is not the case. For these tokenizers, when decoding a sequence of token
IDs to UTF-8 bytes we need to use the token ID -> encoded bytes mappings
provided by the `vocab` map and decode them.

Aside from supporting more tokenizers, using the vocab for decoding is faster
because it involves a single hash map lookup per token, instead of a recursive
splitting of the token ID until we get a sequence of token IDs corresponding to
individual bytes.

In the process, various uses of `str` and `String` that refer to encoded UTF-8
byte sequences were replaced with the `EncodedBytes` and `EncodedByteSlice`
aliases to make it clearer what the content is.

[1] https://huggingface.co/HuggingFaceTB/SmolLM-360M-Instruct